+++ cli/test/smokehouse/core-tests.js
@@ -66,6 +66,8 @@ import sourceMaps from './test-definitions/source-maps.js';
 import timing from './test-definitions/timing.js';
 import trustedTypesDirectivePresent from './test-definitions/trusted-types-directive-present.js';
 import trustedTypesDirectiveMissingDirective from './test-definitions/trusted-types-missing-directives.js';
+import llmsTxt from './test-definitions/llms-txt.js';
+import webmcp from './test-definitions/webmcp.js';
 
 /** @type {ReadonlyArray<Smokehouse.TestDfn>} */
 const smokeTests = [
@@ -96,6 +98,7 @@ const smokeTests = [
   lanternSetTimeout,
   lanternXhr,
   legacyJavascript,
+  llmsTxt,
   metricsDebugger,
   metricsDelayedFcp,
   metricsDelayedLcp,
@@ -131,6 +134,7 @@ const smokeTests = [
   timing,
   trustedTypesDirectivePresent,
   trustedTypesDirectiveMissingDirective,
+  webmcp,
 ];
 
 export default smokeTests;